backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-api-python-client==2.130.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
SQLAlchemy==2.0.32
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1 
pydantic-settings==2.10.1
firebase-admin==6.5.0
fastmcp>=0.4.0